Improve Java Opts Coherence
Previously, the way that JAVA_OPTS were handled was a bit inconsistent. First,
JAVA_OPTS from the environment could be used to affect the memory calculator,
but JAVA_OPTS from configuration could not due to the fact that they were
assembled on the command line after the calculator had run.  In addition, any
changes to the environment's JAVA_OPTS would require a restage as those values
were written into the command line instead of just being passed through the
command stream.

This change attempts to improve the consistency of how JAVA_OPTS are handled
by doing two things.  First, the environment's JAVA_OPTS are never processed
as part of staging.  Instead, a $JAVA_OPTS is appended (or not depending on
configuration) to the JAVA_OPTS contributed to the buildpack.  This solves the
restage issue.  Second, it declares the JAVA_OPTS as the very first part of
the execution command, before the memory calculator.  This solves the issue of
ensuring that both environment and configured JAVA_OPTS are available to the
memory calculator at startup.

[#423][#424]